I accidentally deleted some fxml pages
I added 2 back from the best of memory but they aren't linked to the
code. Also don't rename the localkedro folder because it gets rid of the
project file for some reason. Just create all classes and fxml in the
LocalKedro/src. I know this is trivial but just a reminder