Added type and interface location and event
The location interface is an easy way to put the location of the event
in without using the full address so a vague location can be given
initially and maybe later clarified. Event takes location as a parameter
adn is stored that way. Type is like a party, concert etc that is preset