Added controller classes and real authentication for the user
I added the user and event controller classes as well as authentication
for the user. If you aren't in the user file you can't log in. I added a
way around it in the user.java file. Log in with your first last names
and then your abc123 as the password