Added an rsvp object and redid the user object
each user has a list of rsvp objects that contain the name of the event.
I'm going to add another constructor so the events can have a list of
usernames going to the event. The user object checks in a user folder
and checks the file names of the users. If it matches then it reads the
file into the program. Also added another controller